Use shutil.move instead of copy to reduce IO
According to the docs here: https://docs.python.org/3/library/shutil.html, shutil.move is basically shutil.copy(src, dst) and then os.remove(src), but "if the destination is on the current filesystem, then os.rename() is used". 
In this case shutil.move would reduce the overhead of copying and removing the source file.